move sql and sqlkv backends compatibility tests
