+++ pkg/storage/unified/testing/storage_backend_test.go
@@ -7,11 +7,7 @@ import (
 	badger "github.com/dgraph-io/badger/v4"
 	"github.com/stretchr/testify/require"
 
-	"github.com/grafana/grafana/pkg/infra/db"
-	"github.com/grafana/grafana/pkg/setting"
 	"github.com/grafana/grafana/pkg/storage/unified/resource"
-	sqldb "github.com/grafana/grafana/pkg/storage/unified/sql/db"
-	"github.com/grafana/grafana/pkg/storage/unified/sql/db/dbimpl"
 )
 
 func TestBadgerKVStorageBackend(t *testing.T) {
@@ -41,24 +37,8 @@ func TestBadgerKVStorageBackend(t *testing.T) {
 }
 
 func TestSQLKVStorageBackend(t *testing.T) {
-	newBackendFunc := func(ctx context.Context) (resource.StorageBackend, sqldb.DB) {
-		dbstore := db.InitTestDB(t)
-		eDB, err := dbimpl.ProvideResourceDB(dbstore, setting.NewCfg(), nil)
-		require.NoError(t, err)
-		kv, err := resource.NewSQLKV(eDB)
-		require.NoError(t, err)
-		kvOpts := resource.KVBackendOptions{
-			KvStore: kv,
-		}
-		backend, err := resource.NewKVStorageBackend(kvOpts)
-		require.NoError(t, err)
-		db, err := eDB.Init(ctx)
-		require.NoError(t, err)
-		return backend, db
-	}
-
 	RunStorageBackendTest(t, func(ctx context.Context) resource.StorageBackend {
-		backend, _ := newBackendFunc(ctx)
+		backend, _ := NewTestSqlKvBackend(t, ctx)
 		return backend
 	}, &TestOptions{
 		NSPrefix: "sqlkvstorage-test",
@@ -76,13 +56,6 @@ func TestSQLKVStorageBackend(t *testing.T) {
 			TestGetResourceLastImportTime: true,
 			TestOptimisticLocking:         true,
 			TestKeyPathGeneration:         true,
-		},
-	})
-
-	RunSQLStorageBackendCompatibilityTest(t, newBackendFunc, &TestOptions{
-		NSPrefix: "sqlkvstorage-compatibility-test",
-		SkipTests: map[string]bool{
-			TestKeyPathGeneration: true,
 		},
 	})
 }
